+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/pylapp/miso-ios/releases/comapre/1.0.0..develop)
+
+### Added
+
+- `FoxyRough` theme
+
 ## [1.0.0](https://github.com/pylapp/miso-ios/releases/tag/1.0.0) - 2026-09-16
 
 ### Changed